prefect/prefect_env_requirements.txt
boto3==1.13.9
botocore==1.16.26
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
croniter==0.3.34
dask==2.23.0
distributed==2.23.0
docker==4.3.1
docutils==0.15.2
HeapDict==1.0.1
idna==2.10
jmespath==0.10.0
marshmallow==3.7.1
marshmallow-oneofschema==2.0.1
msgpack==1.0.0
mypy-extensions==0.4.3
natsort==7.0.1
newsapi-python==0.2.6
numpy==1.19.1
pandas==0.25.3
pendulum==2.1.2
prefect==0.13.3
psutil==5.7.2
python-box==4.2.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.15.0
sortedcontainers==2.2.2
tabulate==0.8.7
tblib==1.7.0
text-unidecode==1.3
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
urllib3==1.25.10
websocket-client==0.57.0
zict==2.0.0
